move QueryResponse type back to query-service
