+++ containers/tefca-viewer/src/app/constants.ts
@@ -1,14 +1,3 @@
-import {
-  Patient,
-  Observation,
-  DiagnosticReport,
-  Condition,
-  Encounter,
-  Medication,
-  MedicationAdministration,
-  MedicationRequest,
-} from "fhir/r4";
-
 /**
  * The use cases that can be used in the app
  */
@@ -43,17 +32,3 @@ export type FHIR_SERVERS = (typeof FhirServers)[number];
  */
 export const RequestSources = ["api", "viewer"] as const;
 export type REQUEST_SOURCES = (typeof RequestSources)[number];
-
-/**
- * The query response when the request source is from the Viewer UI.
- */
-export type QueryResponse = {
-  Patient?: Patient[];
-  Observation?: Observation[];
-  DiagnosticReport?: DiagnosticReport[];
-  Condition?: Condition[];
-  Encounter?: Encounter[];
-  Medication?: Medication[];
-  MedicationAdministration?: MedicationAdministration[];
-  MedicationRequest?: MedicationRequest[];
-};
