+++ containers/tefca-viewer/src/app/query-service.ts
@@ -1,8 +1,32 @@
 "use server";
 import fetch from "node-fetch";
+import {
+  Patient,
+  Observation,
+  DiagnosticReport,
+  Condition,
+  Encounter,
+  Medication,
+  MedicationAdministration,
+  MedicationRequest,
+} from "fhir/r4";
 
 import FHIRClient from "./fhir-servers";
-import { USE_CASES, FHIR_SERVERS, QueryResponse } from "./constants";
+import { USE_CASES, FHIR_SERVERS } from "./constants";
+
+/**
+ * The query response when the request source is from the Viewer UI.
+ */
+export type QueryResponse = {
+  Patient?: Patient[];
+  Observation?: Observation[];
+  DiagnosticReport?: DiagnosticReport[];
+  Condition?: Condition[];
+  Encounter?: Encounter[];
+  Medication?: Medication[];
+  MedicationAdministration?: MedicationAdministration[];
+  MedicationRequest?: MedicationRequest[];
+};
 
 export type UseCaseQueryRequest = {
   use_case: USE_CASES;
@@ -17,7 +41,7 @@ const UseCaseQueryMap: {
   [key in USE_CASES]: (
     patientId: string,
     fhirClient: FHIRClient,
-    queryResponse: QueryResponse,
+    queryResponse: QueryResponse
   ) => Promise<QueryResponse>;
 } = {
   "social-determinants": socialDeterminantsQuery,
@@ -42,7 +66,7 @@ export type UseCaseQueryResponse = Awaited<ReturnType<typeof UseCaseQuery>>;
 async function patientQuery(
   request: UseCaseQueryRequest,
   fhirClient: FHIRClient,
-  queryResponse: QueryResponse,
+  queryResponse: QueryResponse
 ): Promise<void> {
   // Query for patient
   let query = "Patient?";
@@ -67,8 +91,8 @@ async function patientQuery(
       `Patient search failed. Status: ${
         response.status
       } \n Body: ${await response.text} \n Headers: ${JSON.stringify(
-        response.headers.raw(),
-      )}`,
+        response.headers.raw()
+      )}`
     );
   }
   queryResponse = await parseFhirSearch(response, queryResponse);
@@ -83,7 +107,7 @@ async function patientQuery(
  */
 export async function UseCaseQuery(
   request: UseCaseQueryRequest,
-  queryResponse: QueryResponse = {},
+  queryResponse: QueryResponse = {}
 ): Promise<QueryResponse> {
   const fhirClient = new FHIRClient(request.fhir_server);
 
@@ -112,7 +136,7 @@ export async function UseCaseQuery(
 async function socialDeterminantsQuery(
   patientId: string,
   fhirClient: FHIRClient,
-  queryResponse: QueryResponse,
+  queryResponse: QueryResponse
 ): Promise<QueryResponse> {
   const query = `/Observation?subject=${patientId}&category=social-history`;
   const response = await fhirClient.get(query);
@@ -129,7 +153,7 @@ async function socialDeterminantsQuery(
 async function newbornScreeningQuery(
   patientId: string,
   fhirClient: FHIRClient,
-  queryResponse: QueryResponse,
+  queryResponse: QueryResponse
 ): Promise<QueryResponse> {
   const loincs: Array<string> = [
     "73700-7",
@@ -161,7 +185,7 @@ async function newbornScreeningQuery(
 async function syphilisQuery(
   patientId: string,
   fhirClient: FHIRClient,
-  queryResponse: QueryResponse,
+  queryResponse: QueryResponse
 ): Promise<QueryResponse> {
   const loincs: Array<string> = ["LP70657-9", "98212-4", "53605-2"];
   const snomed: Array<string> = ["76272004", "186847001"];
@@ -219,7 +243,7 @@ async function syphilisQuery(
 async function gonorrheaQuery(
   patientId: string,
   fhirClient: FHIRClient,
-  queryResponse: QueryResponse,
+  queryResponse: QueryResponse
 ): Promise<QueryResponse> {
   const loincs: Array<string> = [
     "24111-7", // Neisseria gonorrhoeae DNA [Presence] in Specimen by NAA with probe detection
@@ -291,7 +315,7 @@ async function gonorrheaQuery(
 async function chlamydiaQuery(
   patientId: string,
   fhirClient: FHIRClient,
-  queryResponse: QueryResponse,
+  queryResponse: QueryResponse
 ): Promise<QueryResponse> {
   const loincs: Array<string> = [
     "24111-7", // Neisseria gonorrhoeae DNA [Presence] in Specimen by NAA with probe detection
@@ -365,7 +389,7 @@ async function chlamydiaQuery(
 async function cancerQuery(
   patientId: string,
   fhirClient: FHIRClient,
-  queryResponse: QueryResponse,
+  queryResponse: QueryResponse
 ): Promise<QueryResponse> {
   const snomed: Array<string> = ["92814006"];
   const rxnorm: Array<string> = ["828265"]; // drug codes from NLM/NIH RxNorm
@@ -401,7 +425,7 @@ async function cancerQuery(
  */
 async function parseFhirSearch(
   response: fetch.Response | Array<fetch.Response>,
-  queryResponse: QueryResponse = {},
+  queryResponse: QueryResponse = {}
 ): Promise<QueryResponse> {
   let resourceArray: any[] = [];
